Setting Up New Settings Assets
This is incredibly incomplete, as I couldn't really do much more than this. I would not pull this until ready to work with it.

I set up the new settings menu background and buttons. Got started on setting up the popups, but atm if you click into one of the settings options you wont be able to leave as no back button is programmed in.
Additionally I did my best with the dropdowns and sliders, but at some point it stopped cooperating with me as I think it lost the transforms. I was going about it a non-standard way as to not mess with the preexisting components and it might be easier to just reset up the sliders and dropdowns with the new assets. I did get the check mark working though.