Chance PolishedAnims Implemented
Updated Chance FBX with polish Sprint, Walk, and Fear anims. Updated the animation settings to help reflect the animations display the original animation as much as possible. Increased the speed of the new fear anim from 2 to 4 to help reduce the feet sliding. Implemented the new Sprint and Walk anim (with sound). Placed the talk anim in their AC. Made transitions from the fear anim to the other anims to try and get rid of the sliding immediately after you try to move from the fear anim but it did not solve the problem.