Lighting/Fog adjustments
Lighting and fog adjustments. Adjusted the fog color transparency to fade in and out and changed the color to make them alil darker. Alot of lighting work.
-Adjusted lighting for nightstand lamp
-(Nightstand lamp is off for morning and afternoon the light that hits it is from the windows.)
-Turned light setting of for bush particles to get rid of flickering
-Adjusted the window light placements in RC
-Fixed light burn in RC
-Fixed awkward rock shading in zone 2 afternoon by adding a second directional light with low intensity and mixed setting.
-Fixed lamp burn on wall from shelf light in SH
-Asjusted light for white board in EHQ
I tried to fix the shadow render distance around chance for the grass but couldnt find the solution idk if anyone else wants to take a crack at it? While looking into it it could potentially be a setting for the URP? Maybe the grass has a specific setting off?
-I was able to fix the fog popping